+++ src/index.ts
@@ -1,5 +1,6 @@
 import TalkbackFactory from "./talkback-factory"
 import {DefaultOptions, FallbackMode, RecordMode, Options} from "./options"
+import {logger} from "./logger"
 
 const talkback = (options: Partial<Options>) => {
   return TalkbackFactory.server(options)
@@ -12,6 +13,7 @@ talkback.Options = {
 }
 
 talkback.requestHandler = (options: Partial<Options>) => TalkbackFactory.requestHandler(options)
+talkback._logger = logger
 
 export default talkback
 module.exports = talkback