+++ src/types.ts
@@ -1,6 +1,7 @@
 import {FallbackMode, Options, RecordMode} from "./options"
 import TalkbackServer from "./server"
 import RequestHandler from "./request-handler"
+import Logger from "./logger"
 
 export {}
 
@@ -52,4 +53,8 @@ export interface Talkback extends TalkbackBase {
   }
 
   requestHandler(options: Partial<Options>): RequestHandler
+
+  _logger: {
+    log: Logger
+  }
 }