+++ test/summary.spec.ts
@@ -2,8 +2,9 @@ import Summary from "../src/summary"
 import * as td from "testdouble"
 import {DefaultOptions} from "../src/options"
 import Tape from "../src/tape"
+import {logger} from "../src/logger"
 
-let log: Function;
+let logInfo: Function;
 
 const opts = {
   ...DefaultOptions,
@@ -12,7 +13,7 @@ const opts = {
 
 describe("Summary", () => {
   beforeEach(() => {
-    log = td.replace(console, "log")
+    logInfo = td.replace(logger.log, "info")
   })
 
   afterEach(() => {
@@ -20,20 +21,13 @@ describe("Summary", () => {
   })
 
   describe("#print", () => {
-    it("print the server name", () => {
-      const summary = new Summary([], opts)
-      summary.print()
-
-      td.verify(log(td.matchers.contains(opts.name)))
-    })
-
     it("prints nothing when there are no new tapes and no unused tapes", () => {
       const summary = new Summary([], opts)
 
       summary.print()
 
-      td.verify(log(td.matchers.contains("New")), {times: 0})
-      td.verify(log(td.matchers.contains("Unused")), {times: 0})
+      td.verify(logInfo(td.matchers.contains("New")), {times: 0})
+      td.verify(logInfo(td.matchers.contains("Unused")), {times: 0})
     })
 
     it("prints the path of new tapes", () => {
@@ -45,9 +39,9 @@ describe("Summary", () => {
 
       summary.print()
 
-      td.verify(log(td.matchers.contains("path1")))
-      td.verify(log(td.matchers.contains("path2")), {times: 0})
-      td.verify(log(td.matchers.contains("path3")))
+      td.verify(logInfo(td.matchers.contains("path1")))
+      td.verify(logInfo(td.matchers.contains("path2")), {times: 0})
+      td.verify(logInfo(td.matchers.contains("path3")))
     })
 
     it("prints the path of unused tapes", () => {
@@ -59,9 +53,9 @@ describe("Summary", () => {
 
       summary.print()
 
-      td.verify(log(td.matchers.contains("path1")))
-      td.verify(log(td.matchers.contains("path2")), {times: 0})
-      td.verify(log(td.matchers.contains("path3")))
+      td.verify(logInfo(td.matchers.contains("path1")))
+      td.verify(logInfo(td.matchers.contains("path2")), {times: 0})
+      td.verify(logInfo(td.matchers.contains("path3")))
     })
   })
 })